review req: KataConfig.spec.logLevel made optional
Leaving it required would apparently make it necessary to bump the version
of the custom resource (CRD) which seems to cause enough inconvenience not
to be worth the trouble.

A slight optimisation added, too, to prevent ContainerRuntimeConfig
creation when it's not needed, namely in the very common use case of a user
who doesn't care about log level and never sets it.